projects/brownie-exam/pyproject.toml
[project]
name = "brownie-exam"
version = "0.1.0"
description = "For testing Brownie framework"
authors = [
    {name = "hlibbc",email = "hlib,.master@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "eth-brownie (>=1.20.7,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
